Lab9 - Implement Application Security/shopease/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
requests
streamlit
